graphql_schema.js 추가
graphql_schema.js 추가
- 따로 스키마를 관리하기 위해 js 파일 추가
